daemon: start splitting snaps op tests out of api_test.go
this splits /v2/snaps GET and POST tests out of api_test.go
to api_snaps_test.go and api_sideload_n_try_test.go

api.go itself isn't split yet similarly but to clean test writing
two changes were made

* introduction of snapInstruction.dispatchForMany used in postSnaps
  as well
* changed trySnap signature

some imports in api_snaps_test.go are commented out atm,
they were copied from api_test.go and most of them will likely
come back soon
