+++ daemon/api_test.go
@@ -22,29 +22,21 @@ package daemon
 import (
 	"bytes"
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"mime/multipart"
 	"net/http"
 	"net/http/httptest"
-	"net/url"
 	"os"
 	"path/filepath"
-	"regexp"
 	"strings"
 	"time"
 
 	"gopkg.in/check.v1"
 
 	"github.com/snapcore/snapd/arch"
-	"github.com/snapcore/snapd/asserts"
-	"github.com/snapcore/snapd/asserts/assertstest"
 	"github.com/snapcore/snapd/client"
 	"github.com/snapcore/snapd/dirs"
-	"github.com/snapcore/snapd/overlord/assertstate"
-	"github.com/snapcore/snapd/overlord/assertstate/assertstatetest"
 	"github.com/snapcore/snapd/overlord/auth"
 	"github.com/snapcore/snapd/overlord/healthstate"
 	"github.com/snapcore/snapd/overlord/snapstate"
@@ -854,366 +846,6 @@ func (s *apiSuite) TestUserFromRequestHeaderValidUser(c *check.C) {
 	c.Check(user, check.DeepEquals, expectedUser)
 }
 
-func (s *apiSuite) TestSnapsInfoOnePerIntegration(c *check.C) {
-	s.checkSnapInfoOnePerIntegration(c, false, nil)
-}
-
-func (s *apiSuite) TestSnapsInfoOnePerIntegrationSome(c *check.C) {
-	s.checkSnapInfoOnePerIntegration(c, false, []string{"foo", "baz"})
-}
-
-func (s *apiSuite) TestSnapsInfoOnePerIntegrationAll(c *check.C) {
-	s.checkSnapInfoOnePerIntegration(c, true, nil)
-}
-
-func (s *apiSuite) TestSnapsInfoOnePerIntegrationAllSome(c *check.C) {
-	s.checkSnapInfoOnePerIntegration(c, true, []string{"foo", "baz"})
-}
-
-func snapList(rawSnaps interface{}) []map[string]interface{} {
-	snaps := make([]map[string]interface{}, len(rawSnaps.([]*json.RawMessage)))
-	for i, raw := range rawSnaps.([]*json.RawMessage) {
-		err := json.Unmarshal([]byte(*raw), &snaps[i])
-		if err != nil {
-			panic(err)
-		}
-	}
-	return snaps
-}
-
-// XXX this should go away
-var SnapList = snapList
-
-func (s *apiSuite) checkSnapInfoOnePerIntegration(c *check.C, all bool, names []string) {
-	d := s.daemon(c)
-
-	type tsnap struct {
-		name   string
-		dev    string
-		ver    string
-		rev    int
-		active bool
-
-		wanted bool
-	}
-
-	tsnaps := []tsnap{
-		{name: "foo", dev: "bar", ver: "v0.9", rev: 1},
-		{name: "foo", dev: "bar", ver: "v1", rev: 5, active: true},
-		{name: "bar", dev: "baz", ver: "v2", rev: 10, active: true},
-		{name: "baz", dev: "qux", ver: "v3", rev: 15, active: true},
-		{name: "qux", dev: "mip", ver: "v4", rev: 20, active: true},
-	}
-	numExpected := 0
-
-	for _, snp := range tsnaps {
-		if all || snp.active {
-			if len(names) == 0 {
-				numExpected++
-				snp.wanted = true
-			}
-			for _, n := range names {
-				if snp.name == n {
-					numExpected++
-					snp.wanted = true
-					break
-				}
-			}
-		}
-		s.mkInstalledInState(c, d, snp.name, snp.dev, snp.ver, snap.R(snp.rev), snp.active, "")
-	}
-
-	q := url.Values{}
-	if all {
-		q.Set("select", "all")
-	}
-	if len(names) > 0 {
-		q.Set("snaps", strings.Join(names, ","))
-	}
-	req, err := http.NewRequest("GET", "/v2/snaps?"+q.Encode(), nil)
-	c.Assert(err, check.IsNil)
-
-	rsp, ok := getSnapsInfo(snapsCmd, req, nil).(*resp)
-	c.Assert(ok, check.Equals, true)
-
-	c.Check(rsp.Type, check.Equals, ResponseTypeSync)
-	c.Check(rsp.Status, check.Equals, 200)
-	c.Check(rsp.Result, check.NotNil)
-
-	snaps := snapList(rsp.Result)
-	c.Check(snaps, check.HasLen, numExpected)
-
-	for _, s := range tsnaps {
-		if !((all || s.active) && s.wanted) {
-			continue
-		}
-		var got map[string]interface{}
-		for _, got = range snaps {
-			if got["name"].(string) == s.name && got["revision"].(string) == snap.R(s.rev).String() {
-				break
-			}
-		}
-		c.Check(got["name"], check.Equals, s.name)
-		c.Check(got["version"], check.Equals, s.ver)
-		c.Check(got["revision"], check.Equals, snap.R(s.rev).String())
-		c.Check(got["developer"], check.Equals, s.dev)
-		c.Check(got["confinement"], check.Equals, "strict")
-	}
-}
-
-func (s *apiSuite) TestSnapsInfoOnlyLocal(c *check.C) {
-	d := s.daemon(c)
-
-	s.rsnaps = []*snap.Info{{
-		SideInfo: snap.SideInfo{
-			RealName: "store",
-		},
-		Publisher: snap.StoreAccount{
-			ID:          "foo-id",
-			Username:    "foo",
-			DisplayName: "Foo",
-			Validation:  "unproven",
-		},
-	}}
-	s.mkInstalledInState(c, d, "local", "foo", "v1", snap.R(10), true, "")
-	st := s.d.overlord.State()
-	st.Lock()
-	st.Set("health", map[string]healthstate.HealthState{
-		"local": {Status: healthstate.OkayStatus},
-	})
-	st.Unlock()
-
-	req, err := http.NewRequest("GET", "/v2/snaps?sources=local", nil)
-	c.Assert(err, check.IsNil)
-
-	rsp := getSnapsInfo(snapsCmd, req, nil).(*resp)
-
-	c.Assert(rsp.Sources, check.DeepEquals, []string{"local"})
-
-	snaps := snapList(rsp.Result)
-	c.Assert(snaps, check.HasLen, 1)
-	c.Assert(snaps[0]["name"], check.Equals, "local")
-	c.Check(snaps[0]["health"], check.DeepEquals, map[string]interface{}{
-		"status":    "okay",
-		"revision":  "unset",
-		"timestamp": "0001-01-01T00:00:00Z",
-	})
-}
-
-func (s *apiSuite) TestSnapsInfoAllMixedPublishers(c *check.C) {
-	d := s.daemon(c)
-
-	// the first 'local' is from a 'local' snap
-	s.mkInstalledInState(c, d, "local", "", "v1", snap.R(-1), false, "")
-	s.mkInstalledInState(c, d, "local", "foo", "v2", snap.R(1), false, "")
-	s.mkInstalledInState(c, d, "local", "foo", "v3", snap.R(2), true, "")
-
-	req, err := http.NewRequest("GET", "/v2/snaps?select=all", nil)
-	c.Assert(err, check.IsNil)
-	rsp := getSnapsInfo(snapsCmd, req, nil).(*resp)
-	c.Assert(rsp.Type, check.Equals, ResponseTypeSync)
-
-	snaps := snapList(rsp.Result)
-	c.Assert(snaps, check.HasLen, 3)
-
-	publisher := map[string]interface{}{
-		"id":           "foo-id",
-		"username":     "foo",
-		"display-name": "Foo",
-		"validation":   "unproven",
-	}
-
-	c.Check(snaps[0]["publisher"], check.IsNil)
-	c.Check(snaps[1]["publisher"], check.DeepEquals, publisher)
-	c.Check(snaps[2]["publisher"], check.DeepEquals, publisher)
-}
-
-func (s *apiSuite) TestSnapsInfoAll(c *check.C) {
-	d := s.daemon(c)
-
-	s.mkInstalledInState(c, d, "local", "foo", "v1", snap.R(1), false, "")
-	s.mkInstalledInState(c, d, "local", "foo", "v2", snap.R(2), false, "")
-	s.mkInstalledInState(c, d, "local", "foo", "v3", snap.R(3), true, "")
-	s.mkInstalledInState(c, d, "local_foo", "foo", "v4", snap.R(4), true, "")
-	brokenInfo := s.mkInstalledInState(c, d, "local_bar", "foo", "v5", snap.R(5), true, "")
-	// make sure local_bar is 'broken'
-	err := os.Remove(filepath.Join(brokenInfo.MountDir(), "meta", "snap.yaml"))
-	c.Assert(err, check.IsNil)
-
-	expectedHappy := map[string]bool{
-		"local":     true,
-		"local_foo": true,
-		"local_bar": true,
-	}
-	for _, t := range []struct {
-		q        string
-		numSnaps int
-		typ      ResponseType
-	}{
-		{"?select=enabled", 3, "sync"},
-		{`?select=`, 3, "sync"},
-		{"", 3, "sync"},
-		{"?select=all", 5, "sync"},
-		{"?select=invalid-field", 0, "error"},
-	} {
-		c.Logf("trying: %v", t)
-		req, err := http.NewRequest("GET", fmt.Sprintf("/v2/snaps%s", t.q), nil)
-		c.Assert(err, check.IsNil)
-		rsp := getSnapsInfo(snapsCmd, req, nil).(*resp)
-		c.Assert(rsp.Type, check.Equals, t.typ)
-
-		if rsp.Type != "error" {
-			snaps := snapList(rsp.Result)
-			c.Assert(snaps, check.HasLen, t.numSnaps)
-			seen := map[string]bool{}
-			for _, s := range snaps {
-				seen[s["name"].(string)] = true
-			}
-			c.Assert(seen, check.DeepEquals, expectedHappy)
-		}
-	}
-}
-
-func (s *apiSuite) TestSnapsInfoOnlyStore(c *check.C) {
-	d := s.daemon(c)
-
-	s.suggestedCurrency = "EUR"
-
-	s.rsnaps = []*snap.Info{{
-		SideInfo: snap.SideInfo{
-			RealName: "store",
-		},
-		Publisher: snap.StoreAccount{
-			ID:          "foo-id",
-			Username:    "foo",
-			DisplayName: "Foo",
-			Validation:  "unproven",
-		},
-	}}
-	s.mkInstalledInState(c, d, "local", "foo", "v1", snap.R(10), true, "")
-
-	req, err := http.NewRequest("GET", "/v2/snaps?sources=store", nil)
-	c.Assert(err, check.IsNil)
-
-	rsp := getSnapsInfo(snapsCmd, req, nil).(*resp)
-
-	c.Assert(rsp.Sources, check.DeepEquals, []string{"store"})
-
-	snaps := snapList(rsp.Result)
-	c.Assert(snaps, check.HasLen, 1)
-	c.Assert(snaps[0]["name"], check.Equals, "store")
-	c.Check(snaps[0]["prices"], check.IsNil)
-
-	c.Check(rsp.SuggestedCurrency, check.Equals, "EUR")
-}
-
-func (s *apiSuite) TestSnapsInfoStoreWithAuth(c *check.C) {
-	s.daemon(c)
-
-	state := snapCmd.d.overlord.State()
-	state.Lock()
-	user, err := auth.NewUser(state, "username", "email@test.com", "macaroon", []string{"discharge"})
-	state.Unlock()
-	c.Check(err, check.IsNil)
-
-	req, err := http.NewRequest("GET", "/v2/snaps?sources=store", nil)
-	c.Assert(err, check.IsNil)
-
-	c.Assert(s.user, check.IsNil)
-
-	_ = getSnapsInfo(snapsCmd, req, user).(*resp)
-
-	// ensure user was set
-	c.Assert(s.user, check.DeepEquals, user)
-}
-
-func (s *apiSuite) TestSnapsInfoLocalAndStore(c *check.C) {
-	d := s.daemon(c)
-
-	s.rsnaps = []*snap.Info{{
-		Version: "v42",
-		SideInfo: snap.SideInfo{
-			RealName: "remote",
-		},
-		Publisher: snap.StoreAccount{
-			ID:          "foo-id",
-			Username:    "foo",
-			DisplayName: "Foo",
-			Validation:  "unproven",
-		},
-	}}
-	s.mkInstalledInState(c, d, "local", "foo", "v1", snap.R(10), true, "")
-
-	req, err := http.NewRequest("GET", "/v2/snaps?sources=local,store", nil)
-	c.Assert(err, check.IsNil)
-
-	rsp := getSnapsInfo(snapsCmd, req, nil).(*resp)
-
-	// presence of 'store' in sources bounces request over to /find
-	c.Assert(rsp.Sources, check.DeepEquals, []string{"store"})
-
-	snaps := snapList(rsp.Result)
-	c.Assert(snaps, check.HasLen, 1)
-	c.Check(snaps[0]["version"], check.Equals, "v42")
-
-	// as does a 'q'
-	req, err = http.NewRequest("GET", "/v2/snaps?q=what", nil)
-	c.Assert(err, check.IsNil)
-	rsp = getSnapsInfo(snapsCmd, req, nil).(*resp)
-	snaps = snapList(rsp.Result)
-	c.Assert(snaps, check.HasLen, 1)
-	c.Check(snaps[0]["version"], check.Equals, "v42")
-
-	// otherwise, local only
-	req, err = http.NewRequest("GET", "/v2/snaps", nil)
-	c.Assert(err, check.IsNil)
-	rsp = getSnapsInfo(snapsCmd, req, nil).(*resp)
-	snaps = snapList(rsp.Result)
-	c.Assert(snaps, check.HasLen, 1)
-	c.Check(snaps[0]["version"], check.Equals, "v1")
-}
-
-func (s *apiSuite) TestSnapsInfoDefaultSources(c *check.C) {
-	d := s.daemon(c)
-
-	s.rsnaps = []*snap.Info{{
-		SideInfo: snap.SideInfo{
-			RealName: "remote",
-		},
-		Publisher: snap.StoreAccount{
-			ID:          "foo-id",
-			Username:    "foo",
-			DisplayName: "Foo",
-			Validation:  "unproven",
-		},
-	}}
-	s.mkInstalledInState(c, d, "local", "foo", "v1", snap.R(10), true, "")
-
-	req, err := http.NewRequest("GET", "/v2/snaps", nil)
-	c.Assert(err, check.IsNil)
-
-	rsp := getSnapsInfo(snapsCmd, req, nil).(*resp)
-
-	c.Assert(rsp.Sources, check.DeepEquals, []string{"local"})
-	snaps := snapList(rsp.Result)
-	c.Assert(snaps, check.HasLen, 1)
-}
-
-func (s *apiSuite) TestSnapsInfoFilterRemote(c *check.C) {
-	s.daemon(c)
-
-	s.rsnaps = nil
-
-	req, err := http.NewRequest("GET", "/v2/snaps?q=foo&sources=store", nil)
-	c.Assert(err, check.IsNil)
-
-	rsp := getSnapsInfo(snapsCmd, req, nil).(*resp)
-
-	c.Check(s.storeSearch, check.DeepEquals, store.Search{Query: "foo"})
-
-	c.Assert(rsp.Result, check.NotNil)
-}
-
 func (s *apiSuite) TestPostSnapBadRequest(c *check.C) {
 	buf := bytes.NewBufferString(`hello`)
 	req, err := http.NewRequest("POST", "/v2/snaps/hello-world", buf)
@@ -1427,49 +1059,6 @@ func (s *apiSuite) TestPostSnapCohortIncompat(c *check.C) {
 	}
 }
 
-func (s *apiSuite) TestPostSnapVerifyMultiSnapInstruction(c *check.C) {
-	s.daemonWithOverlordMock(c)
-
-	buf := strings.NewReader(`{"action": "install","snaps":["ubuntu-core"]}`)
-	req, err := http.NewRequest("POST", "/v2/snaps", buf)
-	c.Assert(err, check.IsNil)
-	req.Header.Set("Content-Type", "application/json")
-
-	rsp := postSnaps(snapsCmd, req, nil).(*resp)
-
-	c.Check(rsp.Type, check.Equals, ResponseTypeError)
-	c.Check(rsp.Status, check.Equals, 400)
-	c.Check(rsp.Result.(*errorResult).Message, testutil.Contains, `cannot install "ubuntu-core", please use "core" instead`)
-}
-
-func (s *apiSuite) TestPostSnapsNoWeirdses(c *check.C) {
-	s.daemonWithOverlordMock(c)
-
-	// one could add more actions here ... 🤷
-	for _, action := range []string{"install", "refresh", "remove"} {
-		for weird, v := range map[string]string{
-			"channel":      `"beta"`,
-			"revision":     `"1"`,
-			"devmode":      "true",
-			"jailmode":     "true",
-			"cohort-key":   `"what"`,
-			"leave-cohort": "true",
-			"purge":        "true",
-		} {
-			buf := strings.NewReader(fmt.Sprintf(`{"action": "%s","snaps":["foo","bar"], "%s": %s}`, action, weird, v))
-			req, err := http.NewRequest("POST", "/v2/snaps", buf)
-			c.Assert(err, check.IsNil)
-			req.Header.Set("Content-Type", "application/json")
-
-			rsp := postSnaps(snapsCmd, req, nil).(*resp)
-
-			c.Check(rsp.Type, check.Equals, ResponseTypeError)
-			c.Check(rsp.Status, check.Equals, 400)
-			c.Check(rsp.Result.(*errorResult).Message, testutil.Contains, `unsupported option provided for multi-snap operation`)
-		}
-	}
-}
-
 func (s *apiSuite) TestPostSnapSetsUser(c *check.C) {
 	d := s.daemon(c)
 	ensureStateSoon = func(st *state.State) {}
@@ -1542,566 +1131,6 @@ func (s *apiSuite) TestPostSnapEnableDisableSwitchRevision(c *check.C) {
 		c.Check(rsp.Status, check.Equals, 400)
 		c.Check(rsp.Result.(*errorResult).Message, testutil.Contains, "takes no revision")
 	}
-}
-
-var sideLoadBodyWithoutDevMode = "" +
-	"----hello--\r\n" +
-	"Content-Disposition: form-data; name=\"snap\"; filename=\"x\"\r\n" +
-	"\r\n" +
-	"xyzzy\r\n" +
-	"----hello--\r\n" +
-	"Content-Disposition: form-data; name=\"dangerous\"\r\n" +
-	"\r\n" +
-	"true\r\n" +
-	"----hello--\r\n" +
-	"Content-Disposition: form-data; name=\"snap-path\"\r\n" +
-	"\r\n" +
-	"a/b/local.snap\r\n" +
-	"----hello--\r\n"
-
-func (s *apiSuite) TestSideloadSnapOnNonDevModeDistro(c *check.C) {
-	// try a multipart/form-data upload
-	body := sideLoadBodyWithoutDevMode
-	head := map[string]string{"Content-Type": "multipart/thing; boundary=--hello--"}
-	chgSummary := s.sideloadCheck(c, body, head, "local", snapstate.Flags{RemoveSnapPath: true})
-	c.Check(chgSummary, check.Equals, `Install "local" snap from file "a/b/local.snap"`)
-}
-
-func (s *apiSuite) TestSideloadSnapOnDevModeDistro(c *check.C) {
-	// try a multipart/form-data upload
-	body := sideLoadBodyWithoutDevMode
-	head := map[string]string{"Content-Type": "multipart/thing; boundary=--hello--"}
-	restore := sandbox.MockForceDevMode(true)
-	defer restore()
-	flags := snapstate.Flags{RemoveSnapPath: true}
-	chgSummary := s.sideloadCheck(c, body, head, "local", flags)
-	c.Check(chgSummary, check.Equals, `Install "local" snap from file "a/b/local.snap"`)
-}
-
-func (s *apiSuite) TestSideloadSnapDevMode(c *check.C) {
-	body := "" +
-		"----hello--\r\n" +
-		"Content-Disposition: form-data; name=\"snap\"; filename=\"x\"\r\n" +
-		"\r\n" +
-		"xyzzy\r\n" +
-		"----hello--\r\n" +
-		"Content-Disposition: form-data; name=\"devmode\"\r\n" +
-		"\r\n" +
-		"true\r\n" +
-		"----hello--\r\n"
-	head := map[string]string{"Content-Type": "multipart/thing; boundary=--hello--"}
-	// try a multipart/form-data upload
-	flags := snapstate.Flags{RemoveSnapPath: true}
-	flags.DevMode = true
-	chgSummary := s.sideloadCheck(c, body, head, "local", flags)
-	c.Check(chgSummary, check.Equals, `Install "local" snap from file "x"`)
-}
-
-func (s *apiSuite) TestSideloadSnapJailMode(c *check.C) {
-	body := "" +
-		"----hello--\r\n" +
-		"Content-Disposition: form-data; name=\"snap\"; filename=\"x\"\r\n" +
-		"\r\n" +
-		"xyzzy\r\n" +
-		"----hello--\r\n" +
-		"Content-Disposition: form-data; name=\"jailmode\"\r\n" +
-		"\r\n" +
-		"true\r\n" +
-		"----hello--\r\n" +
-		"Content-Disposition: form-data; name=\"dangerous\"\r\n" +
-		"\r\n" +
-		"true\r\n" +
-		"----hello--\r\n"
-	head := map[string]string{"Content-Type": "multipart/thing; boundary=--hello--"}
-	// try a multipart/form-data upload
-	flags := snapstate.Flags{JailMode: true, RemoveSnapPath: true}
-	chgSummary := s.sideloadCheck(c, body, head, "local", flags)
-	c.Check(chgSummary, check.Equals, `Install "local" snap from file "x"`)
-}
-
-func (s *apiSuite) sideloadCheck(c *check.C, content string, head map[string]string, expectedInstanceName string, expectedFlags snapstate.Flags) string {
-	d := s.daemonWithFakeSnapManager(c)
-
-	soon := 0
-	ensureStateSoon = func(st *state.State) {
-		soon++
-		ensureStateSoonImpl(st)
-	}
-
-	c.Assert(expectedInstanceName != "", check.Equals, true, check.Commentf("expected instance name must be set"))
-	mockedName, _ := snap.SplitInstanceName(expectedInstanceName)
-
-	// setup done
-	installQueue := []string{}
-	unsafeReadSnapInfo = func(path string) (*snap.Info, error) {
-		return &snap.Info{SuggestedName: mockedName}, nil
-	}
-
-	snapstateInstall = func(ctx context.Context, s *state.State, name string, opts *snapstate.RevisionOptions, userID int, flags snapstate.Flags) (*state.TaskSet, error) {
-		// NOTE: ubuntu-core is not installed in developer mode
-		c.Check(flags, check.Equals, snapstate.Flags{})
-		installQueue = append(installQueue, name)
-
-		t := s.NewTask("fake-install-snap", "Doing a fake install")
-		return state.NewTaskSet(t), nil
-	}
-
-	snapstateInstallPath = func(s *state.State, si *snap.SideInfo, path, name, channel string, flags snapstate.Flags) (*state.TaskSet, *snap.Info, error) {
-		c.Check(flags, check.DeepEquals, expectedFlags)
-
-		c.Check(path, testutil.FileEquals, "xyzzy")
-
-		c.Check(name, check.Equals, expectedInstanceName)
-
-		installQueue = append(installQueue, si.RealName+"::"+path)
-		t := s.NewTask("fake-install-snap", "Doing a fake install")
-		return state.NewTaskSet(t), &snap.Info{SuggestedName: name}, nil
-	}
-
-	buf := bytes.NewBufferString(content)
-	req, err := http.NewRequest("POST", "/v2/snaps", buf)
-	c.Assert(err, check.IsNil)
-	for k, v := range head {
-		req.Header.Set(k, v)
-	}
-
-	rsp := postSnaps(snapsCmd, req, nil).(*resp)
-	c.Assert(rsp.Type, check.Equals, ResponseTypeAsync)
-	n := 1
-	c.Assert(installQueue, check.HasLen, n)
-	c.Check(installQueue[n-1], check.Matches, "local::.*/"+regexp.QuoteMeta(dirs.LocalInstallBlobTempPrefix)+".*")
-
-	st := d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-	chg := st.Change(rsp.Change)
-	c.Assert(chg, check.NotNil)
-
-	c.Check(soon, check.Equals, 1)
-
-	c.Assert(chg.Tasks(), check.HasLen, n)
-
-	st.Unlock()
-	s.waitTrivialChange(c, chg)
-	st.Lock()
-
-	c.Check(chg.Kind(), check.Equals, "install-snap")
-	var names []string
-	err = chg.Get("snap-names", &names)
-	c.Assert(err, check.IsNil)
-	c.Check(names, check.DeepEquals, []string{expectedInstanceName})
-	var apiData map[string]interface{}
-	err = chg.Get("api-data", &apiData)
-	c.Assert(err, check.IsNil)
-	c.Check(apiData, check.DeepEquals, map[string]interface{}{
-		"snap-name": expectedInstanceName,
-	})
-
-	return chg.Summary()
-}
-
-func (s *apiSuite) TestSideloadSnapJailModeAndDevmode(c *check.C) {
-	body := "" +
-		"----hello--\r\n" +
-		"Content-Disposition: form-data; name=\"snap\"; filename=\"x\"\r\n" +
-		"\r\n" +
-		"xyzzy\r\n" +
-		"----hello--\r\n" +
-		"Content-Disposition: form-data; name=\"jailmode\"\r\n" +
-		"\r\n" +
-		"true\r\n" +
-		"----hello--\r\n" +
-		"Content-Disposition: form-data; name=\"devmode\"\r\n" +
-		"\r\n" +
-		"true\r\n" +
-		"----hello--\r\n"
-	s.daemonWithOverlordMock(c)
-
-	req, err := http.NewRequest("POST", "/v2/snaps", bytes.NewBufferString(body))
-	c.Assert(err, check.IsNil)
-	req.Header.Set("Content-Type", "multipart/thing; boundary=--hello--")
-
-	rsp := postSnaps(snapsCmd, req, nil).(*resp)
-	c.Assert(rsp.Type, check.Equals, ResponseTypeError)
-	c.Check(rsp.Result.(*errorResult).Message, check.Equals, "cannot use devmode and jailmode flags together")
-}
-
-func (s *apiSuite) TestSideloadSnapJailModeInDevModeOS(c *check.C) {
-	body := "" +
-		"----hello--\r\n" +
-		"Content-Disposition: form-data; name=\"snap\"; filename=\"x\"\r\n" +
-		"\r\n" +
-		"xyzzy\r\n" +
-		"----hello--\r\n" +
-		"Content-Disposition: form-data; name=\"jailmode\"\r\n" +
-		"\r\n" +
-		"true\r\n" +
-		"----hello--\r\n"
-	s.daemonWithOverlordMock(c)
-
-	req, err := http.NewRequest("POST", "/v2/snaps", bytes.NewBufferString(body))
-	c.Assert(err, check.IsNil)
-	req.Header.Set("Content-Type", "multipart/thing; boundary=--hello--")
-
-	restore := sandbox.MockForceDevMode(true)
-	defer restore()
-
-	rsp := postSnaps(snapsCmd, req, nil).(*resp)
-	c.Assert(rsp.Type, check.Equals, ResponseTypeError)
-	c.Check(rsp.Result.(*errorResult).Message, check.Equals, "this system cannot honour the jailmode flag")
-}
-
-func (s *apiSuite) TestLocalInstallSnapDeriveSideInfo(c *check.C) {
-	d := s.daemonWithOverlordMock(c)
-	// add the assertions first
-	st := d.overlord.State()
-
-	dev1Acct := assertstest.NewAccount(s.StoreSigning, "devel1", nil, "")
-
-	snapDecl, err := s.StoreSigning.Sign(asserts.SnapDeclarationType, map[string]interface{}{
-		"series":       "16",
-		"snap-id":      "x-id",
-		"snap-name":    "x",
-		"publisher-id": dev1Acct.AccountID(),
-		"timestamp":    time.Now().Format(time.RFC3339),
-	}, nil, "")
-	c.Assert(err, check.IsNil)
-
-	snapRev, err := s.StoreSigning.Sign(asserts.SnapRevisionType, map[string]interface{}{
-		"snap-sha3-384": "YK0GWATaZf09g_fvspYPqm_qtaiqf-KjaNj5uMEQCjQpuXWPjqQbeBINL5H_A0Lo",
-		"snap-size":     "5",
-		"snap-id":       "x-id",
-		"snap-revision": "41",
-		"developer-id":  dev1Acct.AccountID(),
-		"timestamp":     time.Now().Format(time.RFC3339),
-	}, nil, "")
-	c.Assert(err, check.IsNil)
-
-	func() {
-		st.Lock()
-		defer st.Unlock()
-		assertstatetest.AddMany(st, s.StoreSigning.StoreAccountKey(""), dev1Acct, snapDecl, snapRev)
-	}()
-
-	body := "" +
-		"----hello--\r\n" +
-		"Content-Disposition: form-data; name=\"snap\"; filename=\"x.snap\"\r\n" +
-		"\r\n" +
-		"xyzzy\r\n" +
-		"----hello--\r\n"
-	req, err := http.NewRequest("POST", "/v2/snaps", bytes.NewBufferString(body))
-	c.Assert(err, check.IsNil)
-	req.Header.Set("Content-Type", "multipart/thing; boundary=--hello--")
-
-	snapstateInstallPath = func(s *state.State, si *snap.SideInfo, path, name, channel string, flags snapstate.Flags) (*state.TaskSet, *snap.Info, error) {
-		c.Check(flags, check.Equals, snapstate.Flags{RemoveSnapPath: true})
-		c.Check(si, check.DeepEquals, &snap.SideInfo{
-			RealName: "x",
-			SnapID:   "x-id",
-			Revision: snap.R(41),
-		})
-
-		return state.NewTaskSet(), &snap.Info{SuggestedName: "x"}, nil
-	}
-
-	rsp := postSnaps(snapsCmd, req, nil).(*resp)
-	c.Assert(rsp.Type, check.Equals, ResponseTypeAsync)
-
-	st.Lock()
-	defer st.Unlock()
-	chg := st.Change(rsp.Change)
-	c.Assert(chg, check.NotNil)
-	c.Check(chg.Summary(), check.Equals, `Install "x" snap from file "x.snap"`)
-	var names []string
-	err = chg.Get("snap-names", &names)
-	c.Assert(err, check.IsNil)
-	c.Check(names, check.DeepEquals, []string{"x"})
-	var apiData map[string]interface{}
-	err = chg.Get("api-data", &apiData)
-	c.Assert(err, check.IsNil)
-	c.Check(apiData, check.DeepEquals, map[string]interface{}{
-		"snap-name": "x",
-	})
-}
-
-func (s *apiSuite) TestSideloadSnapNoSignaturesDangerOff(c *check.C) {
-	body := "" +
-		"----hello--\r\n" +
-		"Content-Disposition: form-data; name=\"snap\"; filename=\"x\"\r\n" +
-		"\r\n" +
-		"xyzzy\r\n" +
-		"----hello--\r\n"
-	s.daemonWithOverlordMock(c)
-
-	req, err := http.NewRequest("POST", "/v2/snaps", bytes.NewBufferString(body))
-	c.Assert(err, check.IsNil)
-	req.Header.Set("Content-Type", "multipart/thing; boundary=--hello--")
-
-	// this is the prefix used for tempfiles for sideloading
-	glob := filepath.Join(os.TempDir(), "snapd-sideload-pkg-*")
-	glbBefore, _ := filepath.Glob(glob)
-	rsp := postSnaps(snapsCmd, req, nil).(*resp)
-	c.Assert(rsp.Type, check.Equals, ResponseTypeError)
-	c.Check(rsp.Result.(*errorResult).Message, check.Equals, `cannot find signatures with metadata for snap "x"`)
-	glbAfter, _ := filepath.Glob(glob)
-	c.Check(len(glbBefore), check.Equals, len(glbAfter))
-}
-
-func (s *apiSuite) TestSideloadSnapNotValidFormFile(c *check.C) {
-	newTestDaemon(c)
-
-	// try a multipart/form-data upload with missing "name"
-	content := "" +
-		"----hello--\r\n" +
-		"Content-Disposition: form-data; filename=\"x\"\r\n" +
-		"\r\n" +
-		"xyzzy\r\n" +
-		"----hello--\r\n"
-	head := map[string]string{"Content-Type": "multipart/thing; boundary=--hello--"}
-
-	buf := bytes.NewBufferString(content)
-	req, err := http.NewRequest("POST", "/v2/snaps", buf)
-	c.Assert(err, check.IsNil)
-	for k, v := range head {
-		req.Header.Set(k, v)
-	}
-
-	rsp := postSnaps(snapsCmd, req, nil).(*resp)
-	c.Assert(rsp.Type, check.Equals, ResponseTypeError)
-	c.Assert(rsp.Result.(*errorResult).Message, check.Matches, `cannot find "snap" file field in provided multipart/form-data payload`)
-}
-
-func (s *apiSuite) TestSideloadSnapChangeConflict(c *check.C) {
-	body := "" +
-		"----hello--\r\n" +
-		"Content-Disposition: form-data; name=\"snap\"; filename=\"x\"\r\n" +
-		"\r\n" +
-		"xyzzy\r\n" +
-		"----hello--\r\n" +
-		"Content-Disposition: form-data; name=\"dangerous\"\r\n" +
-		"\r\n" +
-		"true\r\n" +
-		"----hello--\r\n"
-	s.daemonWithOverlordMock(c)
-
-	unsafeReadSnapInfo = func(path string) (*snap.Info, error) {
-		return &snap.Info{SuggestedName: "foo"}, nil
-	}
-
-	snapstateInstallPath = func(s *state.State, si *snap.SideInfo, path, name, channel string, flags snapstate.Flags) (*state.TaskSet, *snap.Info, error) {
-		return nil, nil, &snapstate.ChangeConflictError{Snap: "foo"}
-	}
-
-	req, err := http.NewRequest("POST", "/v2/snaps", bytes.NewBufferString(body))
-	c.Assert(err, check.IsNil)
-	req.Header.Set("Content-Type", "multipart/thing; boundary=--hello--")
-
-	rsp := postSnaps(snapsCmd, req, nil).(*resp)
-	c.Assert(rsp.Type, check.Equals, ResponseTypeError)
-	c.Check(rsp.Result.(*errorResult).Kind, check.Equals, client.ErrorKindSnapChangeConflict)
-}
-
-func (s *apiSuite) TestSideloadSnapInstanceName(c *check.C) {
-	// try a multipart/form-data upload
-	body := sideLoadBodyWithoutDevMode +
-		"Content-Disposition: form-data; name=\"name\"\r\n" +
-		"\r\n" +
-		"local_instance\r\n" +
-		"----hello--\r\n"
-	head := map[string]string{"Content-Type": "multipart/thing; boundary=--hello--"}
-	chgSummary := s.sideloadCheck(c, body, head, "local_instance", snapstate.Flags{RemoveSnapPath: true})
-	c.Check(chgSummary, check.Equals, `Install "local_instance" snap from file "a/b/local.snap"`)
-}
-
-func (s *apiSuite) TestSideloadSnapInstanceNameNoKey(c *check.C) {
-	// try a multipart/form-data upload
-	body := sideLoadBodyWithoutDevMode +
-		"Content-Disposition: form-data; name=\"name\"\r\n" +
-		"\r\n" +
-		"local\r\n" +
-		"----hello--\r\n"
-	head := map[string]string{"Content-Type": "multipart/thing; boundary=--hello--"}
-	chgSummary := s.sideloadCheck(c, body, head, "local", snapstate.Flags{RemoveSnapPath: true})
-	c.Check(chgSummary, check.Equals, `Install "local" snap from file "a/b/local.snap"`)
-}
-
-func (s *apiSuite) TestSideloadSnapInstanceNameMismatch(c *check.C) {
-	s.daemonWithFakeSnapManager(c)
-
-	unsafeReadSnapInfo = func(path string) (*snap.Info, error) {
-		return &snap.Info{SuggestedName: "bar"}, nil
-	}
-
-	body := sideLoadBodyWithoutDevMode +
-		"Content-Disposition: form-data; name=\"name\"\r\n" +
-		"\r\n" +
-		"foo_instance\r\n" +
-		"----hello--\r\n"
-
-	req, err := http.NewRequest("POST", "/v2/snaps", bytes.NewBufferString(body))
-	c.Assert(err, check.IsNil)
-	req.Header.Set("Content-Type", "multipart/thing; boundary=--hello--")
-
-	rsp := postSnaps(snapsCmd, req, nil).(*resp)
-	c.Assert(rsp.Type, check.Equals, ResponseTypeError)
-	c.Check(rsp.Result.(*errorResult).Message, check.Equals, `instance name "foo_instance" does not match snap name "bar"`)
-}
-
-func (s *apiSuite) TestTrySnap(c *check.C) {
-	d := s.daemonWithFakeSnapManager(c)
-
-	var err error
-
-	// mock a try dir
-	tryDir := c.MkDir()
-	snapYaml := filepath.Join(tryDir, "meta", "snap.yaml")
-	err = os.MkdirAll(filepath.Dir(snapYaml), 0755)
-	c.Assert(err, check.IsNil)
-	err = ioutil.WriteFile(snapYaml, []byte("name: foo\nversion: 1.0\n"), 0644)
-	c.Assert(err, check.IsNil)
-
-	reqForFlags := func(f snapstate.Flags) *http.Request {
-		b := "" +
-			"--hello\r\n" +
-			"Content-Disposition: form-data; name=\"action\"\r\n" +
-			"\r\n" +
-			"try\r\n" +
-			"--hello\r\n" +
-			"Content-Disposition: form-data; name=\"snap-path\"\r\n" +
-			"\r\n" +
-			tryDir + "\r\n" +
-			"--hello"
-
-		snip := "\r\n" +
-			"Content-Disposition: form-data; name=%q\r\n" +
-			"\r\n" +
-			"true\r\n" +
-			"--hello"
-
-		if f.DevMode {
-			b += fmt.Sprintf(snip, "devmode")
-		}
-		if f.JailMode {
-			b += fmt.Sprintf(snip, "jailmode")
-		}
-		if f.Classic {
-			b += fmt.Sprintf(snip, "classic")
-		}
-		b += "--\r\n"
-
-		req, err := http.NewRequest("POST", "/v2/snaps", bytes.NewBufferString(b))
-		c.Assert(err, check.IsNil)
-		req.Header.Set("Content-Type", "multipart/thing; boundary=hello")
-
-		return req
-	}
-
-	st := d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-
-	for _, t := range []struct {
-		flags snapstate.Flags
-		desc  string
-	}{
-		{snapstate.Flags{}, "core; -"},
-		{snapstate.Flags{DevMode: true}, "core; devmode"},
-		{snapstate.Flags{JailMode: true}, "core; jailmode"},
-		{snapstate.Flags{Classic: true}, "core; classic"},
-	} {
-		soon := 0
-		ensureStateSoon = func(st *state.State) {
-			soon++
-			ensureStateSoonImpl(st)
-		}
-
-		tryWasCalled := true
-		snapstateTryPath = func(s *state.State, name, path string, flags snapstate.Flags) (*state.TaskSet, error) {
-			c.Check(flags, check.DeepEquals, t.flags, check.Commentf(t.desc))
-			tryWasCalled = true
-			t := s.NewTask("fake-install-snap", "Doing a fake try")
-			return state.NewTaskSet(t), nil
-		}
-
-		snapstateInstall = func(ctx context.Context, s *state.State, name string, opts *snapstate.RevisionOptions, userID int, flags snapstate.Flags) (*state.TaskSet, error) {
-			if name != "core" {
-				c.Check(flags, check.DeepEquals, t.flags, check.Commentf(t.desc))
-			}
-			t := s.NewTask("fake-install-snap", "Doing a fake install")
-			return state.NewTaskSet(t), nil
-		}
-
-		// try the snap (without an installed core)
-		st.Unlock()
-		rsp := postSnaps(snapsCmd, reqForFlags(t.flags), nil).(*resp)
-		st.Lock()
-		c.Assert(rsp.Type, check.Equals, ResponseTypeAsync, check.Commentf(t.desc))
-		c.Assert(tryWasCalled, check.Equals, true, check.Commentf(t.desc))
-
-		chg := st.Change(rsp.Change)
-		c.Assert(chg, check.NotNil, check.Commentf(t.desc))
-
-		c.Assert(chg.Tasks(), check.HasLen, 1, check.Commentf(t.desc))
-
-		st.Unlock()
-		s.waitTrivialChange(c, chg)
-		st.Lock()
-
-		c.Check(chg.Kind(), check.Equals, "try-snap", check.Commentf(t.desc))
-		c.Check(chg.Summary(), check.Equals, fmt.Sprintf(`Try "%s" snap from %s`, "foo", tryDir), check.Commentf(t.desc))
-		var names []string
-		err = chg.Get("snap-names", &names)
-		c.Assert(err, check.IsNil, check.Commentf(t.desc))
-		c.Check(names, check.DeepEquals, []string{"foo"}, check.Commentf(t.desc))
-		var apiData map[string]interface{}
-		err = chg.Get("api-data", &apiData)
-		c.Assert(err, check.IsNil, check.Commentf(t.desc))
-		c.Check(apiData, check.DeepEquals, map[string]interface{}{
-			"snap-name": "foo",
-		}, check.Commentf(t.desc))
-
-		c.Check(soon, check.Equals, 1, check.Commentf(t.desc))
-	}
-}
-
-func (s *apiSuite) TestTrySnapRelative(c *check.C) {
-	req, err := http.NewRequest("POST", "/v2/snaps", nil)
-	c.Assert(err, check.IsNil)
-
-	rsp := trySnap(snapsCmd, req, nil, "relative-path", snapstate.Flags{}).(*resp)
-	c.Assert(rsp.Type, check.Equals, ResponseTypeError)
-	c.Check(rsp.Result.(*errorResult).Message, testutil.Contains, "need an absolute path")
-}
-
-func (s *apiSuite) TestTrySnapNotDir(c *check.C) {
-	req, err := http.NewRequest("POST", "/v2/snaps", nil)
-	c.Assert(err, check.IsNil)
-
-	rsp := trySnap(snapsCmd, req, nil, "/does/not/exist", snapstate.Flags{}).(*resp)
-	c.Assert(rsp.Type, check.Equals, ResponseTypeError)
-	c.Check(rsp.Result.(*errorResult).Message, testutil.Contains, "not a snap directory")
-}
-
-func (s *apiSuite) TestTryChangeConflict(c *check.C) {
-	s.daemonWithOverlordMock(c)
-
-	// mock a try dir
-	tryDir := c.MkDir()
-
-	unsafeReadSnapInfo = func(path string) (*snap.Info, error) {
-		return &snap.Info{SuggestedName: "foo"}, nil
-	}
-
-	snapstateTryPath = func(s *state.State, name, path string, flags snapstate.Flags) (*state.TaskSet, error) {
-		return nil, &snapstate.ChangeConflictError{Snap: "foo"}
-	}
-
-	req, err := http.NewRequest("POST", "/v2/snaps", nil)
-	c.Assert(err, check.IsNil)
-
-	rsp := trySnap(snapsCmd, req, nil, tryDir, snapstate.Flags{}).(*resp)
-	c.Assert(rsp.Type, check.Equals, ResponseTypeError)
-	c.Check(rsp.Result.(*errorResult).Kind, check.Equals, client.ErrorKindSnapChangeConflict)
 }
 
 func (s *apiSuite) TestInstallOnNonDevModeDistro(c *check.C) {
@@ -2497,212 +1526,6 @@ func (s *apiSuite) TestSwitchInstruction(c *check.C) {
 	}
 }
 
-func (s *apiSuite) TestPostSnapOp(c *check.C) {
-	s.testPostSnapsOp(c, "application/json")
-}
-
-func (s *apiSuite) TestPostSnapOpMoreComplexContentType(c *check.C) {
-	s.testPostSnapsOp(c, "application/json; charset=utf-8")
-}
-
-func (s *apiSuite) TestPostSnapOpInvalidCharset(c *check.C) {
-	buf := bytes.NewBufferString(`{"action": "refresh"}`)
-	req, err := http.NewRequest("POST", "/v2/snaps", buf)
-	c.Assert(err, check.IsNil)
-	req.Header.Set("Content-Type", "application/json; charset=iso-8859-1")
-
-	rsp := postSnaps(snapsCmd, req, nil).(*resp)
-	c.Check(rsp.Status, check.Equals, 400)
-	c.Check(rsp.Result.(*errorResult).Message, testutil.Contains, "unknown charset in content type")
-}
-
-func (s *apiSuite) testPostSnapsOp(c *check.C, contentType string) {
-	assertstateRefreshSnapDeclarations = func(*state.State, int) error { return nil }
-	snapstateUpdateMany = func(_ context.Context, s *state.State, names []string, userID int, flags *snapstate.Flags) ([]string, []*state.TaskSet, error) {
-		c.Check(names, check.HasLen, 0)
-		t := s.NewTask("fake-refresh-all", "Refreshing everything")
-		return []string{"fake1", "fake2"}, []*state.TaskSet{state.NewTaskSet(t)}, nil
-	}
-
-	d := s.daemonWithOverlordMock(c)
-
-	buf := bytes.NewBufferString(`{"action": "refresh"}`)
-	req, err := http.NewRequest("POST", "/v2/snaps", buf)
-	c.Assert(err, check.IsNil)
-	req.Header.Set("Content-Type", contentType)
-
-	rsp, ok := postSnaps(snapsCmd, req, nil).(*resp)
-	c.Assert(ok, check.Equals, true)
-	c.Check(rsp.Type, check.Equals, ResponseTypeAsync)
-
-	st := d.overlord.State()
-	st.Lock()
-	defer st.Unlock()
-	chg := st.Change(rsp.Change)
-	c.Check(chg.Summary(), check.Equals, `Refresh snaps "fake1", "fake2"`)
-	var apiData map[string]interface{}
-	c.Check(chg.Get("api-data", &apiData), check.IsNil)
-	c.Check(apiData["snap-names"], check.DeepEquals, []interface{}{"fake1", "fake2"})
-}
-
-func (s *apiSuite) TestRefreshAll(c *check.C) {
-	refreshSnapDecls := false
-	assertstateRefreshSnapDeclarations = func(s *state.State, userID int) error {
-		refreshSnapDecls = true
-		return assertstate.RefreshSnapDeclarations(s, userID)
-	}
-	d := s.daemon(c)
-
-	for _, tst := range []struct {
-		snaps []string
-		msg   string
-	}{
-		{nil, "Refresh all snaps: no updates"},
-		{[]string{"fake"}, `Refresh snap "fake"`},
-		{[]string{"fake1", "fake2"}, `Refresh snaps "fake1", "fake2"`},
-	} {
-		refreshSnapDecls = false
-
-		snapstateUpdateMany = func(_ context.Context, s *state.State, names []string, userID int, flags *snapstate.Flags) ([]string, []*state.TaskSet, error) {
-			c.Check(names, check.HasLen, 0)
-			t := s.NewTask("fake-refresh-all", "Refreshing everything")
-			return tst.snaps, []*state.TaskSet{state.NewTaskSet(t)}, nil
-		}
-
-		inst := &snapInstruction{Action: "refresh"}
-		st := d.overlord.State()
-		st.Lock()
-		res, err := snapUpdateMany(inst, st)
-		st.Unlock()
-		c.Assert(err, check.IsNil)
-		c.Check(res.Summary, check.Equals, tst.msg)
-		c.Check(refreshSnapDecls, check.Equals, true)
-	}
-}
-
-func (s *apiSuite) TestRefreshAllNoChanges(c *check.C) {
-	refreshSnapDecls := false
-	assertstateRefreshSnapDeclarations = func(s *state.State, userID int) error {
-		refreshSnapDecls = true
-		return assertstate.RefreshSnapDeclarations(s, userID)
-	}
-
-	snapstateUpdateMany = func(_ context.Context, s *state.State, names []string, userID int, flags *snapstate.Flags) ([]string, []*state.TaskSet, error) {
-		c.Check(names, check.HasLen, 0)
-		return nil, nil, nil
-	}
-
-	d := s.daemon(c)
-	inst := &snapInstruction{Action: "refresh"}
-	st := d.overlord.State()
-	st.Lock()
-	res, err := snapUpdateMany(inst, st)
-	st.Unlock()
-	c.Assert(err, check.IsNil)
-	c.Check(res.Summary, check.Equals, `Refresh all snaps: no updates`)
-	c.Check(refreshSnapDecls, check.Equals, true)
-}
-
-func (s *apiSuite) TestRefreshMany(c *check.C) {
-	refreshSnapDecls := false
-	assertstateRefreshSnapDeclarations = func(s *state.State, userID int) error {
-		refreshSnapDecls = true
-		return nil
-	}
-
-	snapstateUpdateMany = func(_ context.Context, s *state.State, names []string, userID int, flags *snapstate.Flags) ([]string, []*state.TaskSet, error) {
-		c.Check(names, check.HasLen, 2)
-		t := s.NewTask("fake-refresh-2", "Refreshing two")
-		return names, []*state.TaskSet{state.NewTaskSet(t)}, nil
-	}
-
-	d := s.daemon(c)
-	inst := &snapInstruction{Action: "refresh", Snaps: []string{"foo", "bar"}}
-	st := d.overlord.State()
-	st.Lock()
-	res, err := snapUpdateMany(inst, st)
-	st.Unlock()
-	c.Assert(err, check.IsNil)
-	c.Check(res.Summary, check.Equals, `Refresh snaps "foo", "bar"`)
-	c.Check(res.Affected, check.DeepEquals, inst.Snaps)
-	c.Check(refreshSnapDecls, check.Equals, true)
-}
-
-func (s *apiSuite) TestRefreshMany1(c *check.C) {
-	refreshSnapDecls := false
-	assertstateRefreshSnapDeclarations = func(s *state.State, userID int) error {
-		refreshSnapDecls = true
-		return nil
-	}
-
-	snapstateUpdateMany = func(_ context.Context, s *state.State, names []string, userID int, flags *snapstate.Flags) ([]string, []*state.TaskSet, error) {
-		c.Check(names, check.HasLen, 1)
-		t := s.NewTask("fake-refresh-1", "Refreshing one")
-		return names, []*state.TaskSet{state.NewTaskSet(t)}, nil
-	}
-
-	d := s.daemon(c)
-	inst := &snapInstruction{Action: "refresh", Snaps: []string{"foo"}}
-	st := d.overlord.State()
-	st.Lock()
-	res, err := snapUpdateMany(inst, st)
-	st.Unlock()
-	c.Assert(err, check.IsNil)
-	c.Check(res.Summary, check.Equals, `Refresh snap "foo"`)
-	c.Check(res.Affected, check.DeepEquals, inst.Snaps)
-	c.Check(refreshSnapDecls, check.Equals, true)
-}
-
-func (s *apiSuite) TestInstallMany(c *check.C) {
-	snapstateInstallMany = func(s *state.State, names []string, userID int) ([]string, []*state.TaskSet, error) {
-		c.Check(names, check.HasLen, 2)
-		t := s.NewTask("fake-install-2", "Install two")
-		return names, []*state.TaskSet{state.NewTaskSet(t)}, nil
-	}
-
-	d := s.daemon(c)
-	inst := &snapInstruction{Action: "install", Snaps: []string{"foo", "bar"}}
-	st := d.overlord.State()
-	st.Lock()
-	res, err := snapInstallMany(inst, st)
-	st.Unlock()
-	c.Assert(err, check.IsNil)
-	c.Check(res.Summary, check.Equals, `Install snaps "foo", "bar"`)
-	c.Check(res.Affected, check.DeepEquals, inst.Snaps)
-}
-
-func (s *apiSuite) TestInstallManyEmptyName(c *check.C) {
-	snapstateInstallMany = func(_ *state.State, _ []string, _ int) ([]string, []*state.TaskSet, error) {
-		return nil, nil, errors.New("should not be called")
-	}
-	d := s.daemon(c)
-	inst := &snapInstruction{Action: "install", Snaps: []string{"", "bar"}}
-	st := d.overlord.State()
-	st.Lock()
-	res, err := snapInstallMany(inst, st)
-	st.Unlock()
-	c.Assert(res, check.IsNil)
-	c.Assert(err, check.ErrorMatches, "cannot install snap with empty name")
-}
-
-func (s *apiSuite) TestRemoveMany(c *check.C) {
-	snapstateRemoveMany = func(s *state.State, names []string) ([]string, []*state.TaskSet, error) {
-		c.Check(names, check.HasLen, 2)
-		t := s.NewTask("fake-remove-2", "Remove two")
-		return names, []*state.TaskSet{state.NewTaskSet(t)}, nil
-	}
-
-	d := s.daemon(c)
-	inst := &snapInstruction{Action: "remove", Snaps: []string{"foo", "bar"}}
-	st := d.overlord.State()
-	st.Lock()
-	res, err := snapRemoveMany(inst, st)
-	st.Unlock()
-	c.Assert(err, check.IsNil)
-	c.Check(res.Summary, check.Equals, `Remove snaps "foo", "bar"`)
-	c.Check(res.Affected, check.DeepEquals, inst.Snaps)
-}
-
 func (s *apiSuite) TestInstallFails(c *check.C) {
 	snapstateInstall = func(ctx context.Context, s *state.State, name string, opts *snapstate.RevisionOptions, userID int, flags snapstate.Flags) (*state.TaskSet, error) {
 		t := s.NewTask("fake-install-snap-error", "Install task")
@@ -3017,28 +1840,6 @@ func (s *apiSuite) TestLogsNoServices(c *check.C) {
 	rsp := getLogs(logsCmd, req, nil).(*resp)
 	c.Assert(rsp.Status, check.Equals, 404)
 	c.Assert(rsp.Type, check.Equals, ResponseTypeError)
-}
-
-func (s *apiSuite) TestInstallPathUnaliased(c *check.C) {
-	body := "" +
-		"----hello--\r\n" +
-		"Content-Disposition: form-data; name=\"snap\"; filename=\"x\"\r\n" +
-		"\r\n" +
-		"xyzzy\r\n" +
-		"----hello--\r\n" +
-		"Content-Disposition: form-data; name=\"devmode\"\r\n" +
-		"\r\n" +
-		"true\r\n" +
-		"----hello--\r\n" +
-		"Content-Disposition: form-data; name=\"unaliased\"\r\n" +
-		"\r\n" +
-		"true\r\n" +
-		"----hello--\r\n"
-	head := map[string]string{"Content-Type": "multipart/thing; boundary=--hello--"}
-	// try a multipart/form-data upload
-	flags := snapstate.Flags{Unaliased: true, RemoveSnapPath: true, DevMode: true}
-	chgSummary := s.sideloadCheck(c, body, head, "local", flags)
-	c.Check(chgSummary, check.Equals, `Install "local" snap from file "x"`)
 }
 
 type fakeNetError struct {
