+++ daemon/export_test.go
@@ -29,6 +29,7 @@ import (
 	"github.com/snapcore/snapd/overlord"
 	"github.com/snapcore/snapd/overlord/snapstate"
 	"github.com/snapcore/snapd/overlord/state"
+	"github.com/snapcore/snapd/snap"
 )
 
 var MinLane = minLane
@@ -92,11 +93,67 @@ func MockShutdownTimeout(tm time.Duration) (restore func()) {
 	}
 }
 
+func MockUnsafeReadSnapInfo(mock func(string) (*snap.Info, error)) (restore func()) {
+	oldUnsafeReadSnapInfo := unsafeReadSnapInfo
+	unsafeReadSnapInfo = mock
+	return func() {
+		unsafeReadSnapInfo = oldUnsafeReadSnapInfo
+	}
+}
+
+func MockAssertstateRefreshSnapDeclarations(mock func(*state.State, int) error) (restore func()) {
+	oldAssertstateRefreshSnapDeclarations := assertstateRefreshSnapDeclarations
+	assertstateRefreshSnapDeclarations = mock
+	return func() {
+		assertstateRefreshSnapDeclarations = oldAssertstateRefreshSnapDeclarations
+	}
+}
+
 func MockSnapstateInstall(mock func(context.Context, *state.State, string, *snapstate.RevisionOptions, int, snapstate.Flags) (*state.TaskSet, error)) (restore func()) {
 	oldSnapstateInstall := snapstateInstall
 	snapstateInstall = mock
 	return func() {
 		snapstateInstall = oldSnapstateInstall
+	}
+}
+
+func MockSnapstateInstallPath(mock func(*state.State, *snap.SideInfo, string, string, string, snapstate.Flags) (*state.TaskSet, *snap.Info, error)) (restore func()) {
+	oldSnapstateInstallPath := snapstateInstallPath
+	snapstateInstallPath = mock
+	return func() {
+		snapstateInstallPath = oldSnapstateInstallPath
+	}
+}
+
+func MockSnapstateTryPath(mock func(*state.State, string, string, snapstate.Flags) (*state.TaskSet, error)) (restore func()) {
+	oldSnapstateTryPath := snapstateTryPath
+	snapstateTryPath = mock
+	return func() {
+		snapstateTryPath = oldSnapstateTryPath
+	}
+}
+
+func MockSnapstateInstallMany(mock func(*state.State, []string, int) ([]string, []*state.TaskSet, error)) (restore func()) {
+	oldSnapstateInstallMany := snapstateInstallMany
+	snapstateInstallMany = mock
+	return func() {
+		snapstateInstallMany = oldSnapstateInstallMany
+	}
+}
+
+func MockSnapstateUpdateMany(mock func(context.Context, *state.State, []string, int, *snapstate.Flags) ([]string, []*state.TaskSet, error)) (restore func()) {
+	oldSnapstateUpdateMany := snapstateUpdateMany
+	snapstateUpdateMany = mock
+	return func() {
+		snapstateUpdateMany = oldSnapstateUpdateMany
+	}
+}
+
+func MockSnapstateRemoveMany(mock func(*state.State, []string) ([]string, []*state.TaskSet, error)) (restore func()) {
+	oldSnapstateRemoveMany := snapstateRemoveMany
+	snapstateRemoveMany = mock
+	return func() {
+		snapstateRemoveMany = oldSnapstateRemoveMany
 	}
 }
 
@@ -108,4 +165,8 @@ type (
 
 func (inst *snapInstruction) Dispatch() snapActionFunc {
 	return inst.dispatch()
+}
+
+func (inst *snapInstruction) DispatchForMany() snapsActionFunc {
+	return inst.dispatchForMany()
 }
